Add `CanvasViewController.selectedElements()`.
in the states `selectedElement` and `multipleSelectedElements` there
are references to the selected elements, and it was a common pattern to
use a switch case on the `currentState` just to fetch this values, so
created this function to reduce that duplication.

plus, if no element is selected, it returns nil rather than empty array
to make sure when this return not nil, then the array actually contains
selected elements.