Add support for basic copy/cut/paste actions.
notice that the nil-targeted actions for cut, copy and paste are not
present in `NSResponder` then they don’t override anything.

so they are defined just like:

`func copy(_ sender: Any?)`

then there is a basic copying of concepts that just copy its
`attributedStringValue` which has direct support with the pasteboard.

then when pasting i use the `lastCopyIndex` to slightly-translate the
pasted elements each time.

when pasting elements the previously selected elements will be
deselected and the new elements that were pasted will be selected.

in the future i can implement some custom paste support for `Concept`.
Then it’s possible to use the copied point to locate better the next
pasted concept.