Make `StateManager` a `class` instead of a `struct`.
I did not needed the value semantic in this “object” as one is being
references in the `CanvasViewController`.

I was also getting build errors because of
https://github.com/apple/swift-evolution/blob/master/proposals/0176-enfo
rce-exclusive-access-to-memory.md

So as a quick solution i just turned the struct into a class, the real
value type needed is the currentState value, which in this case is an
Enum.