Fix: flip canvas window
the tests needed to be fixed considering that the clicks are made in an
inverted coordinated system.

For the rest of the test the coordinates work as before, but internally,
before sending the event to the view, the coordinate of those clicks
will be inverted, simulating what the iOS coordinate system will see.

Related to #67
Fixes #71