Reset textView attributes when showing it
fixes: https://github.com/fespinoza/LinkedIdeas/issues/92

the main idea is that using `textStorage.setAttributedString` will make the text storage to get the attributes of the last used attributed string

then when passing an empty NSAttributedString, the `NSTextView.typingAttributes` will still have the value of the last used Attributed String, as the empty String does not have attributes (in range) to override with.

Then in that case (only) is important to reset the typing attributes to not to inherit
the last string attributes.

This is not reset when editing an existing string since I do want to use that string attributes to edit it